pyproject.toml
[project]
name = "taskflows"
version = "0.19.0"
description = ""
authors = [
    {name = "Dan Kelleher",email = "kelleherjdan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "docker (>=7.1.0,<8.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "cloudpickle (>=3.1.1,<4.0.0)",
    "dbus-next (>=0.2.3,<0.3.0)",
    "click (>=8.2.1,<9.0.0)",
    "grafanalib (>=0.7.1,<0.8.0)",
    "textdistance (>=4.6.3,<5.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "anyio (>=4.0.0,<5.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.8.0,<3.0.0)",
    "prometheus-client (>=0.19.0,<1.0.0)",
    "async-lru (>=2.0.5,<3.0.0)",
    "imgkit (>=1.2.3,<2.0.0)",
    "dominate (>=2.9.1,<3.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "xxhash (>=3.5.0,<4.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "toolz (>=1.0.0,<2.0.0)",
    "slack-sdk (>=3.36.0,<4.0.0)",
    "emoji (>=2.14.1,<3.0.0)",
    "prettytable (>=3.16.0,<4.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
    "duckdb (>=1.4.0,<2.0.0)",
    "boto3 (>=1.40.40,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydrive (>=1.3.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "^4.0"
pytest-timeout = "^2.0"
pytest-xdist = "^3.0"
hypothesis = "^6.0"
mypy = "^1.8.0"
types-requests = "*"
httpx = "*"
moto = "^5.1.19"


[tool.poetry.scripts]
tf = 'taskflows.admin.cli:cli'
_start_srv_api = 'taskflows.admin.api:_start_api_cmd'
_run_docker_service = 'taskflows.exec:_run_docker_service'
_deserialize_and_call = 'taskflows.exec:_deserialize_and_call'
_run_function = 'taskflows.exec:_run_function'


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: integration tests requiring systemd",
    "security: security-focused tests",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
